Move tests helpers
Create a standalone tests_helpers.rs file to limit the size of
tests_parser.rs with later additions.
